Create if-else-demo-c-program
if and else are two of the most frequently used conditionals in C/C++, and they enable you to execute zero or one conditional statement among many such dependent conditional statements. We use them in the following ways:

if: This executes the body of bracketed code starting with statement1 if condition evaluates to true.

if (condition) { statement1; ... }

if - else: This executes the body of bracketed code starting with statement1 if condition evaluates to true, or it executes the body of code starting with statement2 if condition evaluates to false. Note that only one of the bracketed code sections will ever be executed.

if (condition) { statement1; ... } else { statement2; ... }

Task

Given a positive integer denoting n, do the following:

If 1 <= n <= 9, print the lowercase English word corresponding to the number (e.g., one for 1, two for 2, etc.). If n > 9, print Greater than 9.

Input Format

Get a positive integer from STDIN

Constraints

0

Output Format

Print the relevant output

Sample Input 0

1
Sample Output 0

one